requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
requests
jupyter
